LB1/requirements.txt
pandas>=1.3.0
scikit-learn>=1.0
xgboost>=1.5.0
catboost>=1.0
tensorflow>=2.8.0
matplotlib>=3.5.0
dvc>=2.10.0
scikit-learn
pandas
numpy
matplotlib
seaborn
pyyaml
joblib
tensorboard
